requirements.txt
adbutils==0.10.1
apkutils2==1.0.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
cigam==0.0.3
colorama==0.4.4
colorlog==4.7.2
comtypes==1.1.8
csscompressor==0.9.5
cssselect==1.1.0
decorator==5.0.7
Deprecated==1.2.12
deprecation==2.1.0
facebook-wda==1.4.0
frida==14.2.15
frida-tools==9.2.4
idna==2.10
javaobj-py3==0.4.2
logzero==1.7.0
lxml==4.6.2
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
progress==1.5
prompt-toolkit==3.0.18
psutil==5.8.0
py==1.10.0
pycryptodome==3.10.1
pyelftools==0.27
PyExecJS==1.5.1
Pygments==2.8.1
pyparsing==2.4.7
pyquery==1.4.3
pysqlcipher3==1.0.4
pywin32==300
requests==2.25.1
retry==0.9.2
selenium==3.141.0
six==1.15.0
sqlcipher==0.1.0
tornado==4.5.3
uiautomator2==2.15.0
urllib3==1.26.3
wcwidth==0.2.5
weditor==0.6.3
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
